+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,13 @@
             </intent-filter>
         </activity>
         <activity android:name=".GreetingActivity" />
+        <activity android:name=".LoginActivity"  android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity android:name=".FireApp"
             android:exported="true">
             <intent-filter>
