+++ app/src/main/java/com/github/freeman/bootcamp/MainActivity.kt
@@ -11,54 +11,9 @@ import com.firebase.ui.auth.AuthUI
 import com.google.firebase.auth.FirebaseAuth
 
 class MainActivity : AppCompatActivity() {
-    final val AUTH_REQUEST_code = 7192
-    lateinit var firebaseAuth: FirebaseAuth
-    lateinit var listener: FirebaseAuth.AuthStateListener
-    lateinit var providers: List<AuthUI.IdpConfig>
-
-    override fun onStart() {
-        super.onStart()
-        firebaseAuth.addAuthStateListener(listener)
-        init()
-
-    }
-
-    override fun onStop() {
-        if (listener != null) {
-            firebaseAuth.addAuthStateListener(listener);
-        }
-        super.onStop()
-    }
-
-    private fun init() {
-        providers = arrayListOf(
-            AuthUI.IdpConfig.GoogleBuilder().build()
-        )
-        firebaseAuth = FirebaseAuth.getInstance()
-        listener = object : FirebaseAuth.AuthStateListener {
-            override fun onAuthStateChanged(p0: FirebaseAuth) {
-                val user = p0.currentUser
-                if (user != null)// Already Login
-                {
-                    //Do something
-                    Toast.makeText(this@MainActivity, "" + user.uid, Toast.LENGTH_SHORT).show()
-                } else {
-                    startActivity(
-                        AuthUI.getInstance()
-                            .createSignInIntentBuilder()
-                            .setAvailableProviders(providers)
-                            .build()
-                    )
-                }
-            }
-
-        }
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        init()
     }
 
     fun greet(view: View) {
